Move liquibase checking to ActualLiquibaseMigrationsChecker class
