+++ core/distributed-task-spring-boot-autoconfigure/src/main/java/com/distributed_task_framework/autoconfigure/validation/CheckActualLiquibaseMigrationsListener.java
@@ -1,0 +1,15 @@
+package com.distributed_task_framework.autoconfigure.validation;
+
+import liquibase.integration.spring.SpringLiquibase;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.event.ContextRefreshedEvent;
+
+public class CheckActualLiquibaseMigrationsListener implements ApplicationListener<ContextRefreshedEvent> {
+
+    @Override
+    public void onApplicationEvent(ContextRefreshedEvent event) {
+        var springLiquibase = event.getApplicationContext().getBean(SpringLiquibase.class);
+        var checker = event.getApplicationContext().getBean(ActualLiquibaseMigrationsChecker.class);
+        checker.check(springLiquibase);
+    }
+}
